Fix a compiler warning.

+++ src/unit_animation.cpp
@@ -877,7 +877,9 @@ void unit_animator::wait_until(int animation_time) const
 	while (SDL_GetTicks() < (unsigned int)end_tick - 20*disp->turbo_speed()) {
 		disp->draw();
 		events::pump();
-		disp->delay(maximum<int>(0,minimum<int>(10,(animation_time - get_animation_time())*disp->turbo_speed()) ));
+		disp->delay(maximum<int>(0,
+			minimum<int>(10, 
+			static_cast<int>((animation_time - get_animation_time()) * disp->turbo_speed()))));
 	}
 	disp->delay(maximum<int>(0,end_tick - SDL_GetTicks() +5));
 	new_animation_frame();
